fix: switch rotate and scale trigger implementation
the 'reparent' implementation seems to have graphical issues, where the
objects don't display at the right position when rotating.
this commit switches it to an implementation where it calculates a
position delta manually instead of leaving it up to godot to handle with
reparenting.
this was actually the first implementation of this but it was broken
because the scale trigger used unmodified code i copied from the rotate
trigger's code (i maybe forgot to modify it?)